Re-track some necessary repos for AOSP 14
In #148 - a WIP PR - a bunch of annoying repos were untracked to see if
they could finally be skipped.  Unfortunately it turns out most of these
have to be restored to make the build succeed.

In addition `bash` history suggests that I went through the same hopes
and process last year.